[INV-67] Add pylint formatter
* Fix existing lint issues
* Add lint to pre-push hooks
* Moved python files to api_src
