+++ git-hook-src/pre-push
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+${BASE_DIR}/server/lint.sh
