+++ server/api_src/db.py
@@ -1,4 +1,5 @@
+"""Database module"""
 from flask_sqlalchemy import SQLAlchemy
 
 
-db = SQLAlchemy()
+DB = SQLAlchemy()
