+++ server/build.sh
@@ -4,6 +4,10 @@ set -e
 
 echo "Building..."
 
+export BASE_DIR=/
+. /venv/bin/activate
+./lint.sh
+
 echo "Compiling Sass into CSS..."
 ./sass/dart-sass/sass ./styles/style.scss ./public/css/styles.css
