+++ server/lint.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+cd ${BASE_DIR}/server && python -m pylint --rcfile .pylintrc api_src
