+++ server/package.json
@@ -11,29 +11,29 @@
     "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
-    "axios": "^0.19.0",
-    "core-js": "^2.6.5",
-    "vue": "^2.6.10",
-    "vue-class-component": "^7.0.2",
-    "vue-property-decorator": "^8.1.0",
-    "vue-resource": "^1.5.1",
-    "vue-router": "^3.0.3",
-    "vuex": "^3.0.1"
+    "axios": "0.19.0",
+    "core-js": "2.6.5",
+    "vue": "2.6.10",
+    "vue-class-component": "7.0.2",
+    "vue-property-decorator": "8.1.0",
+    "vue-resource": "1.5.1",
+    "vue-router": "3.0.3",
+    "vuex": "3.0.1"
   },
   "devDependencies": {
-    "@types/jest": "^23.1.4",
-    "@vue/cli-plugin-babel": "^3.9.0",
-    "@vue/cli-plugin-typescript": "^3.9.0",
-    "@vue/cli-plugin-unit-jest": "^3.9.0",
-    "@vue/cli-service": "^3.9.0",
+    "@types/jest": "23.1.4",
+    "@vue/cli-plugin-babel": "3.9.0",
+    "@vue/cli-plugin-typescript": "3.9.0",
+    "@vue/cli-plugin-unit-jest": "3.9.0",
+    "@vue/cli-service": "3.9.0",
     "@vue/test-utils": "1.0.0-beta.29",
     "babel-core": "7.0.0-bridge.0",
-    "lint-staged": "^8.1.5",
-    "sass": "^1.18.0",
-    "sass-loader": "^7.1.0",
-    "ts-jest": "^23.0.0",
-    "typescript": "^3.4.3",
-    "vue-template-compiler": "^2.6.10"
+    "lint-staged": "8.1.5",
+    "sass": "1.18.0",
+    "sass-loader": "7.1.0",
+    "ts-jest": "23.0.0",
+    "typescript": "3.4.3",
+    "vue-template-compiler": "2.6.10"
   },
   "postcss": {
     "plugins": {
