+++ server/requirements.txt
@@ -1,4 +1,5 @@
 apispec==2.0.2
+astroid==2.2.5
 certifi==2019.6.16
 chardet==3.0.4
 Click==7.0
@@ -6,14 +7,20 @@ Flask==1.1.1
 flask-apispec==0.8.1
 Flask-SQLAlchemy==2.4.0
 idna==2.8
+isort==4.3.21
 itsdangerous==1.1.0
 Jinja2==2.10.1
+lazy-object-proxy==1.4.1
 MarkupSafe==1.1.1
 marshmallow==2.19.5
+mccabe==0.6.1
 psycopg2==2.8.3
+pylint==2.3.1
 requests==2.22.0
 six==1.12.0
 SQLAlchemy==1.3.5
+typed-ast==1.4.0
 urllib3==1.25.3
 webargs==5.4.0
 Werkzeug==0.15.4
+wrapt==1.11.2
